requirements.txt
aiomysql==0.2.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
httpx==0.27.0
h2==4.1.0
cryptography==42.0.5
mangum==0.17.0
boto3==1.34.110
starlette==0.37.2
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
motor==3.4.0
pymongo==4.7.2
requests==2.31.0